BookingForm fix.
